fix(cluster): #4356 read the gateway self-id from config, not the bootstrap race
#4351 shipped and did nothing. Setting `gateway_preferred_instance_id:
mac-book-release` on both nodes and restarting mac-book produced:

    19:15:25.925  GATEWAY-LEASE: Claude launch skipped — singleton lease held elsewhere
    19:15:26.543  [cluster] runtime bootstrapped instance_id="mac-book-release"

The preferred node acquired its lease 0.6s *before* cluster bootstrap published
`SELF_INSTANCE_ID`. `resolve_self_instance_id_without_config()` therefore returned
its `hostname-pid` fallback, `self_is_preferred()` was false, and mac-book took the
non-preferred branch: it never called `register_gateway_waiter`, so
`discord_gateway.waiting_providers` stayed `[]`, so the holder never saw a reason
to yield. The handover could not happen at all.

The race was already documented one function over: `wait_for_self_instance_id` says
it exists for "callers that race with cluster bootstrap and would otherwise pick up
the hostname+PID fallback."

`gateway_preference()` already loads the whole config, and `cluster.instance_id` is
the same value bootstrap publishes. Reading it directly removes the race instead of
timing around it; only an unset (`auto`) id still has to wait for bootstrap to
derive one. Both call sites are async, and the keepalive resolves once inside its
spawned task.

The seven existing tests could not have caught this: they all hand a self-id to
`resolve_gateway_preference`, a pure function. How the caller obtained that id was
outside the test surface. The two new tests cover that layer with
`SELF_INSTANCE_ID` unset — the exact state at gateway-acquire time — and both fail
if the fix is reverted.

Tests moved to gateway_lease_tests.rs: the file crossed its 700-LoC namespace cap.
